stability fixes

+++ modelserver/src/main/java/app/beamcatcher/modelserver/io/sadrema/out/SadremaSolutionRetrieverRunnable.java
@@ -4,8 +4,12 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
+import java.net.InetAddress;
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
@@ -17,6 +21,7 @@ import app.beamcatcher.modelserver.configuration.Configuration;
 import app.beamcatcher.modelserver.io.WorldSemaphore;
 import app.beamcatcher.modelserver.model.Beam;
 import app.beamcatcher.modelserver.model.Footprint;
+import app.beamcatcher.modelserver.model.Marker;
 import app.beamcatcher.modelserver.model.SADREMAGridCell;
 import app.beamcatcher.modelserver.model.User;
 import app.beamcatcher.modelserver.model.World;
@@ -36,35 +41,62 @@ public class SadremaSolutionRetrieverRunnable implements Runnable {
 
 		logger.info("Starting polling !");
 
+		String allMarkerJSONPrevious = "";
+
 		try {
 
 			do {
+
+				logger.info("Analyzing world...");
+
 				WorldSemaphore.semaphore.acquire();
 
 				final Map<UUID, User> mapUser = WorldSingleton.INSTANCE.getMapUser();
+				final Map<UUID, Marker> mapAllMarker = new HashMap<UUID, Marker>();
 				final Set<UUID> setUUIDUser = mapUser.keySet();
+				final List<UUID> listAllMarkerUUID = new ArrayList<UUID>();
 				Integer totalMarkers = 0;
 				for (UUID userUUID : setUUIDUser) {
 					final User user = mapUser.get(userUUID);
-					totalMarkers = totalMarkers + user.getMapMarker().size();
+					final HashMap<UUID, Marker> mapMarkerForUser = user.getMapMarker();
+					totalMarkers = totalMarkers + mapMarkerForUser.size();
+					final Set<UUID> setUUIDMarkerForUser = mapMarkerForUser.keySet();
+					for (UUID uuidMarker : setUUIDMarkerForUser) {
+						listAllMarkerUUID.add(uuidMarker);
+						final Marker marker = mapMarkerForUser.get(uuidMarker);
+						mapAllMarker.put(uuidMarker, marker);
+					}
 				}
 
 				if (totalMarkers > 0) {
 
-					final StringBuffer markersCSVInput = WorldToMarkersCSVFileMapper.toMarkersCSVFile();
+					List<Marker> listMarker = new ArrayList<Marker>();
+					Collections.sort(listAllMarkerUUID);
+					for (UUID uuidMarker : listAllMarkerUUID) {
+						final Marker marker = mapAllMarker.get(uuidMarker);
+						listMarker.add(marker);
+					}
+					String allMarkerJSON = Util.getJSON(listMarker);
 
-					CSVWriter.writeToFiles(markersCSVInput);
+					if (!allMarkerJSON.equals(allMarkerJSONPrevious)) {
+						logger.info("Markers changed !");
+						logger.info("Invoking sadrema...");
+						final StringBuffer markersCSVInput = WorldToMarkersCSVFileMapper.toMarkersCSVFile();
+						CSVWriter.writeToFiles(markersCSVInput);
+						final File[] files = SadremaHelper.obtainSolution(listMarker);
+						applyBeams(files[1], listMarker);
+						files[0].delete();
+						File input = new File(Configuration.MODEL_SERVER_IO_DIR_SADREMA_OUT_DATA + "/markers.csv");
+						input.delete();
+						allMarkerJSONPrevious = allMarkerJSON;
+					} else {
+						logger.info("No change in markers");
+						logger.info("Sadrema not invoked");
+					}
 
-					final File beamsTXTFile = SadremaHelper.obtainSolution();
-
-					applyBeams(beamsTXTFile);
-
-					beamsTXTFile.delete();
-
-					File input = new File(Configuration.MODEL_SERVER_IO_DIR_SADREMA_OUT_DATA + "/markers.csv");
-
-					input.delete();
-
+				} else {
+					logger.info("No markers");
+					logger.info("Sadrema not invoked");
 				}
 
 				WorldSemaphore.semaphore.release();
@@ -81,8 +113,7 @@ public class SadremaSolutionRetrieverRunnable implements Runnable {
 		}
 	}
 
-	private void applyBeams(final File pFile) {
-		logger.info("Processing beams from Sadrema:");
+	private void applyBeams(final File pFile, final List<Marker> pMarker) {
 		BufferedReader br = null;
 		try {
 			br = new BufferedReader(new FileReader(pFile));
@@ -136,16 +167,6 @@ public class SadremaSolutionRetrieverRunnable implements Runnable {
 				}
 			}
 
-			for (Beam beam : sat1Beam) {
-				logger.info(beam.toString());
-			}
-			for (Beam beam : sat2Beam) {
-				logger.info(beam.toString());
-			}
-			for (Beam beam : sat3Beam) {
-				logger.info(beam.toString());
-			}
-
 			// Apply to model
 
 			// If world is not valid due to being mutated in the meantime, discard...
@@ -157,6 +178,13 @@ public class SadremaSolutionRetrieverRunnable implements Runnable {
 			newWorld.addBeamsToSatellite(Configuration.WORLD_SATELLITE_3_NAME, sat3Beam);
 			if (ModelValidator.isValid(newWorld)) {
 				WorldSingleton.INSTANCE = newWorld;
+				SadremaHelper.doPNG(pMarker, pFile.getParent(), beamsPerSat);
+				final String hostname = InetAddress.getLocalHost().getHostName();
+				final String context = "archive";
+				final String[] split = pFile.getParent().split("/");
+				final String dirname = split[split.length - 1];
+				final String URL = "http://" + hostname + "/" + context + "/" + dirname;
+				logger.info("Results at " + URL);
 			}
 
 		} catch (Exception e) {
